Improve Active Time estimation for Arms Warrior in Cataclysm (#7171)
* adjust Arms GCDs to ignore haste

* refactor APL code to allow use within an Analyzer.

some features are not supported in this context. notably: lookaheads.

this will error if an APL tries to use a lookahead in an Analyzer context.

looking through usage, it appears that only Devastation and Protection Paladin
use lookaheads still. Paladin is disabled and Devastation's APL is disabled,
so the feature could probably be removed.

consider this one step on the road to full Analyzer-ization

* use an APL to estimate unavoidable downtime for Cata Arms

Arms is a downtime spec in Cata: you do not have enough cooldowns
or Rage to fill every GCD, so some are unavoidably empty.

this uses an APL-based approach to attempt to identify situations
where there isn't an appropriate spell to cast. such periods are then
added to the `AlwaysBeCasting` as "uptime" regions, boosting overall uptime

* update Arms guide text

* changelog entries

* test fix

* Update src/common/SPELLS/classic/warrior.ts

Co-authored-by: Jazminite <65823478+jazminite@users.noreply.github.com>

---------

Co-authored-by: Jazminite <65823478+jazminite@users.noreply.github.com>