add support for San'layn Blood DK  (#7191)
* update breath of fire text again

* add support for San'layn Blood DK

some of this probably also works for Unholy DK, but afaik 0 people play that and I'm not familiar enough with the differences to know.

The most complex piece of this is the modification to the `Haste` module to allow updating the value of a haste buff. Gift of the San'layn
increases the haste % on Essence of the Blood Queen from 1% to 3% for its duration. This means that we need to change the haste value for
the buff mid-combat. AFAIK this is totally unprecedented. This is implemented by basically removing the old haste and then applying the new one.