Caster/Healer version of the new Foundation Downtime timeline (#7288)
* WIP caster timeline downtime

* implement healer version of downtime section

this involved fixing a bug in the healing active time for `AlwaysBeCastingHealing` where (some?) accesses would see the wrong list of healing abilities.

* use `<Suggestions>` component instead

still not using `suggestion` on the analyzers but maybe i should fix the API instead of duplicating it...

* adjust scrollbar css for windows

* use actual suggestion api (shocker)

* fix display issues with suggestions

* fix lint errors, adjust api

* convert HEALING_ABILITIES_ON_GCD to instance property

this *technically* causes more memory consumption but like...not really, in the grand scheme of things.

several implementations already used it as an instance property, which coincidentally fixed the bug in the default `isHealingAbility` implementation

* changelog entry