Mage: add HYPERTHERMIA_BUFF for heating up. (#7355)
* Mage: add HYPERTHERMIA_BUFF for heating up.

Before only HYPERTHERMIA_TALENT was added.
The missing HYPERTHERMIA_BUFF was causing casts of IB/PF during HYPERTHERMIA to be treated as mistakes (not converting heating up to hot streak!).

* Mage: add HYPERTHERMIA_BUFF for heating up.

Before only HYPERTHERMIA_TALENT was added.
The missing HYPERTHERMIA_BUFF was causing casts of IB/PF during HYPERTHERMIA to be treated as mistakes (not converting heating up to hot streak!).