Use DBC data to generate ability lists for Classic specs (#7413)
* fix wowhead tooltip url for MoP logs

this WAS working until i fixed the thing on the WCL backend that made the logs
properly identify as mists.

* initial pass on dbc ability gen

* small refactor

* mop blood dk

* bump spec versions so CI hits them

* prettier generated spell lists by default

* okay so i used the wrong prettier config?

* lint issues

* switch to version of wow-dbc that has a license

* mop: arms warrior

* improve generated spell naming + consistency

* mop: brewmaster monk

Jab causes some headaches

* mop: windwalker monk

* mop: mistweaver monk

plus some general changes to make the extra-long Xuen duration play nice.

as well as some changes to HealingEfficiency to handle the not-SPELL spells
from gen.

* adjust spellgen to avoid `readonly` problem

tl;dr - adding `satisfies RetailSpell` fixes it.

in order for that to work, some internal fields needed
to be stripped from the data. they aren't part of the public interface
from wow-dbc so wow-dbc should *probably* be stripping them but
it is tricky to do it generically. still, probably better to have wow-dbc
do it statically than wowa

* lint fixes

* arms war changelog

* mop talent support

* timeline tweaks. make abilities bigger

* add icons to generated spells

turns out the file isn't THAT big, at least for classic. only...28mb of json

* normalizer for brewmaster xuen in classic

* add script to update spell lists

* lint issues

* switch to github repo

* fix bug in icon name extraction